app/src/main/java/in/utilitysoftwares/androidtestingapplication/shoppinglistapp/data/local/ShoppingDao.kt
package `in`.utilitysoftwares.androidtestingapplication.shoppinglistapp.data.local

import androidx.lifecycle.LiveData
import androidx.room.*

@Dao
interface ShoppingDao {
    @Insert(onConflict = OnConflictStrategy.REPLACE)
    suspend fun insertShoppingItem(shoppingItem: ShoppingItem)

    @Delete
    suspend fun deleteShoppingItem (shoppingDao: ShoppingItem)

    @Query("SELECT * FROM shopping_items")
    fun observeAllShoppingItems(): LiveData<List<ShoppingItem>>

    @Query("SELECT SUM( price * amount) FROM shopping_items")
    fun observeTotalPrice(): LiveData<Float>
}